# openapi.yaml
openapi: 3.0.3
info:
  title: CruiseFeed API
  version: "1.0.0"
  description: >
    The CruiseFeed API delivers normalized cruise inventory data — cruise lines,
    ships, sailing dates, day-by-day itineraries, ports of call and lead-in
    pricing — as a clean JSON or CSV REST API, refreshed up to daily.


    Built for travel agencies, OTAs, cruise-comparison sites and data
    aggregators who would rather consume one normalized feed than build and
    maintain scrapers. Every record has a stable opaque `id` and one schema across all cruise lines.


    Get an API key and free sample at https://cruisefeed.io.
  termsOfService: https://cruisefeed.io
  contact:
    name: CruiseFeed
    url: https://cruisefeed.io
    email: support@cruisefeed.io
  license:
    name: Proprietary — data licensed per the CruiseFeed subscription terms
    url: https://cruisefeed.io

servers:
  - url: https://api.cruisefeed.io
    description: Production

security:
  - ApiKeyAuth: []

tags:
  - name: Cruises
    description: Search, export and retrieve sailings.
  - name: Ships
    description: Vessel metadata (specs, capacity, build, dimensions).
  - name: History
    description: Price & availability change tracking.
  - name: Reference
    description: Reference lists (cruise lines, ships, ports). Requires an API key.

paths:
  /cruises:
    get:
      tags: [Cruises]
      summary: List & filter cruises
      description: >
        Paginated, filterable list of sailings. The workhorse endpoint —
        combine any of the filters below. Results are deduplicated by default.
      operationId: listCruises
      parameters:
        - { name: cruise_line, in: query, schema: { type: string }, description: "Exact cruise line name (see GET /cruise-lines)." }
        - { name: ship_name, in: query, schema: { type: string }, description: "Exact ship name." }
        - { name: embark_port, in: query, schema: { type: string }, description: "Departure (embarkation) port." }
        - { name: region, in: query, schema: { type: string }, description: "Destination/region, partial match (e.g. Caribbean)." }
        - { name: departure_from, in: query, schema: { type: string, format: date }, description: "Sailings departing on/after (YYYY-MM-DD)." }
        - { name: departure_to, in: query, schema: { type: string, format: date }, description: "Sailings departing on/before (YYYY-MM-DD)." }
        - { name: min_price, in: query, schema: { type: number }, description: "Minimum lead-in price." }
        - { name: max_price, in: query, schema: { type: number }, description: "Maximum lead-in price." }
        - { name: min_nights, in: query, schema: { type: integer }, description: "Minimum nights." }
        - { name: max_nights, in: query, schema: { type: integer }, description: "Maximum nights." }
        - { name: round_trip, in: query, schema: { type: boolean }, description: "Only round-trip (true) or one-way (false)." }
        - { name: dedupe, in: query, schema: { type: boolean, default: true }, description: "Collapse duplicate sailings. Default true." }
        - { name: sort, in: query, schema: { type: string, default: departure_date }, description: "Sort field. Default departure_date." }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 50 }, description: "Page size, 1–500." }
        - { name: offset, in: query, schema: { type: integer, default: 0 }, description: "Rows to skip." }
      responses:
        "200":
          description: A page of sailings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CruiseList" }
        "401":
          $ref: "#/components/responses/Unauthorized"

  /cruises.csv:
    get:
      tags: [Cruises]
      summary: Export cruises as CSV
      description: Same filters as `GET /cruises`, streamed as a CSV file for spreadsheets and warehouse loads.
      operationId: exportCruisesCsv
      parameters:
        - { name: cruise_line, in: query, schema: { type: string } }
        - { name: ship_name, in: query, schema: { type: string } }
        - { name: embark_port, in: query, schema: { type: string } }
        - { name: region, in: query, schema: { type: string } }
        - { name: departure_from, in: query, schema: { type: string, format: date } }
        - { name: departure_to, in: query, schema: { type: string, format: date } }
        - { name: min_price, in: query, schema: { type: number } }
        - { name: max_price, in: query, schema: { type: number } }
        - { name: min_nights, in: query, schema: { type: integer } }
        - { name: max_nights, in: query, schema: { type: integer } }
        - { name: round_trip, in: query, schema: { type: boolean } }
        - { name: dedupe, in: query, schema: { type: boolean, default: true } }
      responses:
        "200":
          description: CSV export.
          content:
            text/csv:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"

  /cruises/{id}:
    get:
      tags: [Cruises]
      summary: Get a single cruise
      description: Fetch one sailing by its opaque id (returned on every record).
      operationId: getCruise
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: "The opaque cruise id." }
      responses:
        "200":
          description: The sailing, enriched with its matched ship.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CruiseDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  /cruises/{id}/history:
    get:
      tags: [History]
      summary: Price & availability history
      description: Point-in-time price snapshots for a sailing, oldest to newest — track how a fare has moved.
      operationId: getCruiseHistory
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, description: "The opaque cruise id." }
      responses:
        "200":
          description: Price history for the sailing.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CruiseHistory" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  /changes:
    get:
      tags: [History]
      summary: Recent price changes
      description: Sailings whose price changed between the two most recent snapshots — ideal for fare-drop alerts.
      operationId: listChanges
      parameters:
        - { name: since, in: query, schema: { type: string, format: date }, description: "Only changes on/after this date (YYYY-MM-DD)." }
        - { name: cruise_line, in: query, schema: { type: string }, description: "Filter by cruise line." }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Recent price changes.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChangePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /ships:
    get:
      tags: [Ships]
      summary: List & filter ships
      description: >
        Vessel metadata — capacity, gross tonnage, length/beam, decks, cabins, builder,
        class, sister ships, flag and more — for 1,500+ ships, sourced from cruisemapper.
      operationId: listShips
      parameters:
        - { name: q, in: query, schema: { type: string }, description: "Case-insensitive ship-name search." }
        - { name: operator, in: query, schema: { type: string }, description: "Filter by operator (partial match)." }
        - { name: flag_state, in: query, schema: { type: string }, description: "Filter by flag state (partial match)." }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: A page of ships.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShipPage" }
        "401":
          $ref: "#/components/responses/Unauthorized"

  /ships/{ship_id}:
    get:
      tags: [Ships]
      summary: Get a single ship
      description: One vessel by its source_id — the IMO number (or a URL slug when IMO is absent).
      operationId: getShip
      parameters:
        - { name: ship_id, in: path, required: true, schema: { type: string }, description: "IMO number, e.g. 9187796." }
      responses:
        "200":
          description: The ship.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ship" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  /cruise-lines:
    get:
      tags: [Reference]
      summary: List cruise line names
      description: Distinct cruise line names in the feed.
      operationId: listCruiseLines
      responses:
        "200":
          description: Cruise line names.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValueList" }
              example: { items: ["AIDA Cruises", "Carnival", "MSC Cruises", "Royal Caribbean", "Viking"], total: 5 }

  /ports:
    get:
      tags: [Reference]
      summary: List departure ports
      description: Distinct embarkation ports in the feed.
      operationId: listPorts
      responses:
        "200":
          description: Port names.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValueList" }

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Your CruiseFeed API key. Get one at https://cruisefeed.io.

  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: unauthorized, message: "Invalid or missing API key." } }
    NotFound:
      description: No matching record.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: not_found, message: "Not found." } }

  schemas:
    Cruise:
      type: object
      description: A single sailing, normalized across all cruise lines.
      properties:
        id: { type: string, example: cru_9f3a2b71c4d5e6f7, description: "Opaque, source-agnostic id; the key for the detail and history endpoints." }
        cruise_line: { type: string, example: MSC Cruises }
        ship_name: { type: string, example: MSC World Europa }
        title: { type: string, example: 7-Night Western Mediterranean }
        departure_date: { type: string, format: date, example: "2026-09-12" }
        return_date: { type: string, format: date, example: "2026-09-19" }
        duration_days: { type: integer, example: 8 }
        nights: { type: integer, example: 7 }
        round_trip: { type: boolean, example: true }
        embark_port: { type: string, example: Barcelona }
        disembark_port: { type: string, example: Barcelona }
        region: { type: string, example: Western Mediterranean }
        price_amount: { type: number, example: 799, description: "Lead-in (lowest) fare; see `fares` for the full per-cabin breakdown." }
        price_currency: { type: string, example: USD }
        price_per_night: { type: number, example: 114.14 }
        fares:
          type: array
          description: Per-cabin-class fare breakdown; empty when the source quotes only a single "from" price.
          items: { $ref: "#/components/schemas/Fare" }
        itinerary:
          type: array
          items: { $ref: "#/components/schemas/ItineraryStop" }
        scraped_at: { type: string, format: date-time, example: "2026-06-28T04:12:00Z" }

    ValueList:
      type: object
      description: Envelope for reference/facet endpoints.
      properties:
        items: { type: array, items: { type: string } }
        total: { type: integer, example: 6 }

    Fare:
      type: object
      description: One cabin-class fare for a sailing.
      properties:
        cabin_class: { type: string, example: Balcony }
        cabin_code: { type: string, nullable: true, example: BB }
        price_amount: { type: number, example: 1149 }
        price_currency: { type: string, example: USD }
        price_per_night: { type: number, example: 164.14 }
        available: { type: boolean, nullable: true, example: true }

    ItineraryStop:
      type: object
      properties:
        seq: { type: integer, example: 2 }
        port: { type: string, example: Marseille }
        port_code: { type: string, nullable: true, example: MRS }
        arrive: { type: string, nullable: true }
        depart: { type: string, nullable: true }
        is_embark: { type: boolean, example: false }
        is_disembark: { type: boolean, example: false }

    CruiseList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Cruise" }
        total: { type: integer, example: 1843, description: "Total matches across all pages." }
        limit: { type: integer, example: 50 }
        offset: { type: integer, example: 0 }

    Ship:
      type: object
      description: Vessel metadata for one ship.
      properties:
        source: { type: string, example: cruisemapper }
        source_id: { type: string, example: "9187796", description: "IMO number (natural key)." }
        imo_number: { type: string, example: "9187796" }
        mmsi_number: { type: string, nullable: true }
        ship_name: { type: string, example: MSC World Europa }
        operator: { type: string, example: MSC Cruises }
        owner: { type: string, example: "MSC Cruises S.A." }
        year_built: { type: integer, example: 2022 }
        last_refurbishment: { type: integer, nullable: true, example: 2024 }
        passengers: { type: string, example: "2210 - 2680", description: "Raw capacity range." }
        passengers_max: { type: integer, example: 6762 }
        crew: { type: integer, example: 2138 }
        passengers_to_space_ratio: { type: number, example: 39 }
        gross_tonnage: { type: integer, example: 215863 }
        length_m: { type: number, example: 333 }
        beam_m: { type: number, example: 47 }
        speed_knots: { type: number, example: 22 }
        decks: { type: integer, example: 22 }
        decks_with_cabins: { type: integer, example: 16 }
        cabins: { type: integer, example: 2626 }
        ship_class: { type: string, example: World }
        builder: { type: string, example: "Chantiers de l'Atlantique" }
        building_cost: { type: string, example: "EUR 1 billion" }
        flag_state: { type: string, example: Malta }
        engines_power: { type: string }
        propulsion_power: { type: string }
        sister_ships: { type: array, items: { type: string } }
        former_names: { type: array, items: { type: string } }
        christened_by: { type: string, nullable: true }
        ship_url: { type: string }
        cover_image: { type: string }
        scraped_at: { type: string, format: date-time }

    ShipPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Ship" }
        total: { type: integer, example: 1500 }
        limit: { type: integer, example: 50 }
        offset: { type: integer, example: 0 }

    CruiseDetail:
      description: A single sailing, enriched with its matched ship (null if unknown).
      allOf:
        - $ref: "#/components/schemas/Cruise"
        - type: object
          properties:
            ship: { $ref: "#/components/schemas/Ship" }

    HistoryPoint:
      type: object
      properties:
        snapshot_date: { type: string, format: date, example: "2026-06-15" }
        price_amount: { type: number, example: 849 }
        price_currency: { type: string, example: USD }
        price_per_night: { type: number, example: 121.29 }
        nights: { type: integer, example: 7 }
        captured_at: { type: string, format: date-time, example: "2026-06-15T04:10:00Z" }

    CruiseHistory:
      type: object
      properties:
        id: { type: string, example: cru_9f3a2b71c4d5e6f7 }
        points:
          type: array
          items: { $ref: "#/components/schemas/HistoryPoint" }

    PriceChange:
      type: object
      properties:
        id: { type: string, example: cru_9f3a2b71c4d5e6f7 }
        cruise_line: { type: string, example: MSC Cruises }
        ship_name: { type: string, example: MSC World Europa }
        departure_date: { type: string, format: date, example: "2026-09-12" }
        old_price: { type: number, example: 849 }
        new_price: { type: number, example: 799 }
        price_currency: { type: string, example: USD }
        old_date: { type: string, format: date, example: "2026-06-15" }
        new_date: { type: string, format: date, example: "2026-06-28" }

    ChangePage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PriceChange" }
        total: { type: integer, example: 42 }
        limit: { type: integer, example: 100 }
        offset: { type: integer, example: 0 }

    Error:
      type: object
      description: "Uniform error envelope: {\"error\": {\"code\", \"message\"}}."
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: unauthorized }
            message: { type: string, example: "Invalid or missing API key." }